added ConceptExercises java Project
ConceptExercises is a work in progress.  Basically an organizational framework for demonstrating the various topics for the class. In essence, there is a switch which governs demonstrations of various topics.  Currently there is support for the various Collections types with no actual demonstrations. More framework to be in place shortly.